backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow>=9.0.0
pillow-heif>=0.10.0
Werkzeug==2.3.7
